Move StatisticsRange to data models
